Factor out client-side-validation.tsx.
